sales_agent/requirements.txt
google-adk
google-cloud-bigquery
google-cloud-aiplatform[agent-engines]
db-dtypes
pandas
python-dotenv
google-auth
